Extract draft-preview serving out of submissions.ts
Phase 3 Wave B, next seam: pulls canPlayDraft, replyWithDraft, and
replyWithStoredDraft — plus the /api/drafts/:slug and
/api/submissions/:token/preview routes — into
apps/api/src/delivery/draft-preview-routes.ts, registered from inside
registerSubmissionRoutes the same way the catalog-routes seam was.

The one remaining external call site (the draft fallback branch in
/api/games/:slug) is unchanged: the destructured canPlayDraft/replyWithDraft
from the new module's handle keep the same names, so submissions.ts calls
them exactly as before. The /api/submissions/:token/shot/:id build-screenshot
route stays put — it shares no code with this cluster (token auth, not the
draft-sharing check) and keeps its Map-sharing relationship with
catalog-routes.ts's gallery-media route untouched.

submissions.ts drops from 5791 to 5529 lines with no behavior change,
including the pre-existing (vestigial) githubClient gate on both moved
routes, carried forward as a plain githubConfigured boolean.
